refactor(users): move controllers to file
